requirements/requirements.txt
opencv-python
mediapipe
fastapi[all]
uvicorn
requests
PyQt5
auto-py-to-exe
win10toast
numpy
openai
periodictable
pygame
qrcode
matplotlib
pandas
plyer
chemparse
dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition
quickchart-io